add order API with ddl auto
