+++ src/main/java/com/DataBaseProject/PCenter/data/Order.java
@@ -1,0 +1,2 @@
+package com.DataBaseProject.PCenter.data;public class Order {
+}
