+++ src/main/java/com/DataBaseProject/PCenter/exception/ResourceNotFoundException.java
@@ -1,0 +1,2 @@
+package com.DataBaseProject.PCenter;public class ResourceNotFoundException {
+}
