+++ src/main/java/com/DataBaseProject/PCenter/repository/OrderProductRepository.java
@@ -1,0 +1,2 @@
+package com.DataBaseProject.PCenter.repository;public interface OrderProductRepository {
+}
