+++ src/main/java/com/DataBaseProject/PCenter/service/OrderProductServiceImpl.java
@@ -1,0 +1,2 @@
+package com.DataBaseProject.PCenter.service;public class OrderProductServiceImpl {
+}
